Update API review migraton to support on demand migration of individual review
