+++ src/dotnet/APIView/CloneAPIViewDB/apiviewdbmigration.yml
@@ -1,6 +1,11 @@
 trigger: none
 pr: none
 
+parameters:
+- name: ReviewId
+  type: string
+  displayName: 'Old Review ID'
+
 variables:
   ProjectDirectory: 'src/dotnet/APIView/CloneAPIViewDB/CloneAPIView'  
   NugetSecurityAnalysisWarningLevel: 'none'
@@ -23,7 +28,7 @@ stages:
               DOTNET_CLI_TELEMETRY_OPTOUT: 1
               DOTNET_MULTILEVEL_LOOKUP: 0
 
-          - script: 'dotnet run'
+          - script: 'CloneAPIViewDB ${{ parameters.ReviewId }}'
             displayName: 'Run Migration'
             workingDirectory: $(ProjectDirectory)
             env:
